新增功能 - 新增查詢使用者不良紀錄類型

+++ src/main/java/tw/edu/ntub/imd/camping/service/UserService.java
@@ -14,7 +14,7 @@ public interface UserService extends BaseService<UserBean, String> {
 
     List<UserBadRecordBean> getBadRecord(String account);
 
-    Integer getComment(String account);
+    Double getComment(String account);
 
     void compensate(String account, CreditCard creditCard);
